Refactor project structure (c.p. functions)

+++ ubtg/ai_gemini.py
@@ -5,7 +5,7 @@ from google import genai
 GEMINI_API_KEY = config.load(config.API_KEYS_PATH)['GEMINI_API']
 
 
-def generate_content(prompt: str | None = None, system_prompt: str | None = None, media: list[str] | None = None, model: str = 'gemini-2.0-flash', temperature: int = 1, api_key: str = GEMINI_API_KEY):
+def generate_content(prompt: str | None = None, system_prompt: str | None = None, media: list[str] | None = None, model: str = 'gemini-2.0-flash', temperature: float = 1, api_key: str = GEMINI_API_KEY):
     if (not prompt) and (not media):
         raise ValueError('No "prompt" or "images" provided')
